backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.14.0
flower==1.2.0
greenlet==2.0.1
humanize==4.4.0
idna==3.4
kombu==5.2.4
prometheus-client==0.15.0
prompt-toolkit==3.0.36
pytz==2022.7
redis==4.4.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.45
sqlparse==0.4.3
tornado==6.2
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
